requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
flask-restx==1.1.0
numpy==1.23.5
pandas==1.5.3
statsmodels==0.13.5
torch==2.0.0
lightgbm==3.3.5
pmdarima==2.0.3
sktime==0.16.1
ipykernel==6.22.0
prophet==1.1.2
waitress==2.1.2
python-dotenv==1.0.0
## The following requirements were added by pip freeze:
aniso8601==9.0.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
click==8.1.3
cmdstanpy==1.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
ephem==4.1.4
executing==1.2.0
filelock==3.10.7
fonttools==4.39.3
hijri-converter==2.2.4
holidays==0.22
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
llvmlite==0.39.1
LunarCalendar==0.0.9
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.56.4
packaging==23.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
PyMeeus==0.5.12
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
